Separate classes and ClassName to different files
