+++ src/Character/Classes/ClassName.ts
@@ -1,0 +1,9 @@
+enum ClassName {
+    Fighter = 'Fighter',
+    Priest = 'Priest',
+    Ranger = 'Ranger',
+    Rogue = 'Rogue',
+    Wizard = 'Wizard',
+}
+
+export default ClassName;
